+++ packages/ergo-engine/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.53"></a>
+## [0.0.53](https://github.com/accordproject/ergo/compare/v0.0.52...v0.0.53) (2018-05-24)
+
+
+
+
+**Note:** Version bump only for package @accordproject/ergo-engine
+
+<a name="0.0.52"></a>
+## [0.0.52](https://github.com/accordproject/ergo/compare/v0.0.48...v0.0.52) (2018-05-24)
+
+
+
+
+**Note:** Version bump only for package @accordproject/ergo-engine
+
 <a name="0.0.50"></a>
 ## [0.0.50](https://github.com/accordproject/ergo/compare/v0.0.49...v0.0.50) (2018-05-24)